PLAYG-20 Fixing destroy errors

+++ osbrain/app.py
@@ -20,12 +20,11 @@ api = Api(app)
 class WorldRes(Resource):
     
     def get(self):
-        name = 'Emptiness' if not World.get_name() else World.get_name()
-        msg = f'Welcome to {name}! Please look at the [Cloud Run/Debug Locally - Detailed] window'
-        return {'msg': msg}, 200 if World.get_name() else 404
+        if not World.get_name(): return {'msg': 'Welcome to Emptiness'}, 404
+        return {'msg': f'Welcome to {World.get_name()}!'}, 200
     
     def put(self):
-        http_code = 201 if not World.get_live_ns() else 200
+        http_code = 201 if not World.get_name() else 200
         data = request.get_json()
         World(data['name'])
         msg = f'Welcome to {data["name"]}!'
@@ -33,20 +32,22 @@ class WorldRes(Resource):
 
     def delete(self):
         name = World.get_name()
-        World.destroy_world()   
+        if not name: return {'msg': 'Nothing to destroy here'}, 404  
+        World.destroy_world() 
         return {'msg': f'{name} was destroyed!'}, 200
 
 
 class PushPullRes(Resource):
     
     def get(self):
+        if not World.get_name(): return {'msg': 'Cannot communicate in vacuum'}, 404
         agent = AgentUtils('6uv5UhwyXmnStkYRYFdq')
         nick = agent.get_nick()
         ns_name = agent.get_ns_name()
         if agent.check_alive():
             for i in range(3):
                 agent.get_agent_proxy(ns_name).send('main', f'Hello there! This is {nick}! Message #{i}')
-                time.sleep(3)
+                time.sleep(0.5)
             msg = f'{nick} says:- Please look for messages in the console'
         else:
             msg = f'{nick} says:- Looks like I am dead'
